refactor: unify component imports and enhance UI components
