+++ biblio-app/app/welcome.tsx
@@ -3,10 +3,7 @@ import { Pressable, View } from 'react-native';
 import { KeyboardAwareScrollView } from 'react-native-keyboard-controller';
 import { SafeAreaView, useSafeAreaInsets } from 'react-native-safe-area-context';
 import { Button } from '~/components/nativewindui/Button';
-import { FormBlock, FormRow } from '~/components/nativewindui/FormGroup';
-import { ToggleGroup } from '~/components/nativewindui/ToggleGroup';
-import { InputField } from '~/components/nativewindui/InputField';
-import { Icon, Text } from '~/components/ui';
+import { Icon, Text, InputField, FormBlock, FormRow, ToggleGroup } from '~/components/ui';
 import { useAuthStore } from '~/store';
 import { ActivityIndicator } from '~/components/nativewindui/ActivityIndicator';
 import { SCHOOL_ID } from '~/lib/utils';
