+++ biblio-app/components/ui/index.ts
@@ -1,3 +1,8 @@
 export * from './Text';
 export * from './Icon';
 export * from './BaseCard';
+export * from './SearchField';
+export * from './FormGroup';
+export * from './InputField';
+export * from './Stepper';
+export * from './ToggleGroup';
